mejoras para playstore

+++ app/src/main/java/com/example/catjump/domain/model/Cat.kt
@@ -12,7 +12,7 @@ data class Cat(
     val fatness: Float = 0f,  // 0.0 = normal, 1.0 = muy gordo (máx 10 comidas)
     val birdsEaten: Int = 0,  // Contador de pájaros/ratones comidos
     val lives: Int = 3,       // Vidas (ovillos de lana)
-    val invincibilityFrames: Int = 0,  // Frames de invencibilidad después de daño
+    val invincibilityFrames: Float = 0f,  // Frames de invencibilidad restantes (se descuenta por delta-time)
     // Power-ups activos
     val jetpackActive: Boolean = false,
     val jetpackEndTime: Long = 0L,
@@ -24,6 +24,6 @@ data class Cat(
     val centerY: Float get() = y + height / 2
     val bottom: Float get() = y + height
     val right: Float get() = x + width
-    val isInvincible: Boolean get() = invincibilityFrames > 0 || jetpackActive || superJumpActive
+    val isInvincible: Boolean get() = invincibilityFrames > 0f || jetpackActive || superJumpActive
     val hasPowerUp: Boolean get() = jetpackActive || superJumpActive
 }
